requirements.txt
# Core
Django>=1.10,<1.11
psycopg2>=2.7.1

# Heroku
dj-database-url>=0.4.1
gunicorn>=19.6.0


# Storage S3
whitenoise>=3.2

# rest api
djangorestframework>=3.6.2
Markdown>=2.6.8
django-filter>=1.0.2
django-rest-swagger>=2.1.2

# devices
# http://django-osm-field.readthedocs.io/en/latest/
django-osm-field>=0.3.1
uuid>=1.30